Deduplicate datetime/timestamp parsing functions
Move them to `utils.js` and drop duplicates in components
`YmdTimestampHeader` and `TimestampHeader`.
